25.01.22 모듈화 수정 완료
