+++ SalCap/src/main/resources/templates/index.html
@@ -247,7 +247,7 @@ body {
 }
 
     </style>
-    <link rel="stylesheet" href="styles.css">
+    <!--<link rel="stylesheet" href="styles.css">-->
 </head>
 <body>
     <header class="header">
